Migrate underlying abi to `witx2` crate
This migrates the ABI bits and pieces currently in the `witx` crate into
the `witx2` crate. As part of this transition a few updates are made to
bring this in line with the most recent canonical ABI as well:

* Flags are sequential `i32` values
* Return pointer areas of exports are allocated by the callee
* Memory functions have been renamed
* Semantics around handles and drop will be tweaked

Tons of changes here, and wasmtime's bindings generation is still not
compiling. That comes next! Very few tests passing yet.