Sync with wasmtime's new API, Implement new syntax (#11)
* Start updating for the new-api branch

* Get tests to the point they can run again

* Deal with endianness later

* Get exports tests running in some capacity again

* Import initial witx2 crate

* Prep for new parsing AST

* Brace import lists

* Update type syntax

Don't allow anonymous records/enums/flags/unions/etc

* Update function syntax

* Parse functions in resources

* Move around wasi files

* Implement interfaces later

* Migrate underlying abi to `witx2` crate

This migrates the ABI bits and pieces currently in the `witx` crate into
the `witx2` crate. As part of this transition a few updates are made to
bring this in line with the most recent canonical ABI as well:

* Flags are sequential `i32` values
* Return pointer areas of exports are allocated by the callee
* Memory functions have been renamed
* Semantics around handles and drop will be tweaked

Tons of changes here, and wasmtime's bindings generation is still not
compiling. That comes next! Very few tests passing yet.

* Get gen-wasmtime compiling again

* Realloc fixes

* Tweak destructor semantics

* Improve flags in-memory representation

Use u8-u64 as the native representation where possible, and only fall
back onto [i32] when required. Current Rust generators don't support 3
i32 values at this time.

* Use upstream wasmtime

* Remove stray debugging line

* Start adding more tests

* Setup tests for both rust-wasm and wasmtime generators

* Get wasmtime tests running again

* Automatically test checked/unchecked for rust-wasm

* Add more witx tests

Fix an issue with generated code trying to cast `&u32` to `i32` with a
helper function that transparently loads.

* Fix names of variants in lifting/lowering

Also fix `Copy` being applied to enums with `Vec` inside them.

* Fix a typo in size/align calculations

* Add some tests for flags and variants

* Don't use patch to depend on wasmtime

* Remove unnecessary `[patch]`

* Import wasi-snapshot-preview1, test its bindings

Test that we can generate bindings for now, although don't exercise
anything as-of just yet.

* wasmtime: Give preview1 ABI access to raw memory

We can only do it unsafely, though, since preview1 doesn't know what the
pointer is going to be used for.

* Add wasi examples with new syntax

* Remove no longer needed code

* Remove some stray TODO

* Change how tuples are parsed, parse named results

* Implicitly define resources in the type namespace

* Update test expectation

* Deterministically add names in glob imports