Reduce unsafety in wasmtime bindings (#16)
* Reduce unsafety in wasmtime bindings

This commit removes a number of `unsafe` blocks both in the generated
code as well as internally within the bindings library support. This is
mostly just a refactoring of existing functionality but expressed in a
way that we don't have to write `unsafe` in as many places. Additionally
the `region.rs` tests have been re-enabled where appropriate.

* Leverage `AllBytesValid` for some more types

* Update some comments

* Use an upstream method instead of one defined locally