Reimplement code generation.
This commit reimplements code generation such that it uses a custom adapter for
calling from parent into adapted module.

witx-bindgen was used before to generate code; however, the intention of
witx-bindgen code generation is to convert from a language-specific ABI to the
canonical ABI.

This complicated what needed to happen for adapting a call where both sides
were using the canonical ABI, especially when it came to implement support for
variants.

Instead, this uses a simpler adapter that essentially finds lists in the types
being exchanged and ensures the proper marshaling of the list data.