Implement an async mode for wasmtime imports (#19)
This commit implements initial support for async communication with a
wasm module. This implements async support for both calling exports as
well as host-defined imports. The intention of this mode is that
functions can be explicitly configured as either `async` or not, and the
bindings generator will appropriately generate bindings for all functions.

For example if any exported function is asynchronous, then all of them
must be asynchronous because Wasmtime can't mix sync and async calls.
Imports, however, can be mixed-defined as async and sync.

A new test mode was added to ensure that async generation works for all
types. Additionally some wasmtime-specific tests were added about mixing
async/sync calls to ensure that the right bindings are generated.