Consolidate function-local state in function-local structs (#22)
* js: Store function-local state in a separate struct

This commit moves function-local state in the JS function generator into
a separate function-specific struct, avoiding `impl Bindgen for Js`.
This removes the need to reset state on `import`/`export` and makes it a
bit more clean what's called and used where.

* Place function-specific state in a function-local struct

This applies the same change from the JS generator to the Wasmtime
generator as well. This removes the need to reset state between
import/export calls and clearly scopes function-level information to
just one function.

* Do the same as the previous 2 commit for rust-wasm

This has a number of other changes as well such as plumbing `Direction`
around instead of a `bool`. This also refactors the gen-rust-core traits
to be more suitable for direction/lifting/lowering by trying to not bake
those assumptions in so deeply.

* Fix demo build