Reimplement exported resources (#27)
* Reimplement exported resources

This commit goes over the implemented generators and reimplements
exported resource types. This was implemented historically but in the
most recent revision of the canonical ABI various bits and pieces had
changed and this project hadn't been updated yet. This updates all the
various generators to the most recent version of the canonical ABI and
adds tests/etc. Mostly this just uncomments the existing tests and gets
them working.

The main change from before is that handles crossing the abi boundary
are always table indices rather than sometimes indices and sometimes
pointers. The boxing happens internally within the wasm module and the
host is responsible for managing tables. Additionally all host resources
are implicitly reference counted, although hosts don't necessarily have
to expose that to their runtimes. For example the reference counting is
optional in JS with a `FinalizationRegistry` used for cleanup.

* Update the demo

Use a persistent configuration object which makes the render process a
bit easier.