Copy witx files into `test-modules` directory
For now the intention is that the `tests` directory is primarily tests
for code generators, where adding more witx files there is intended to
exercise more cases in code generators, not necessarily having backing
code somewhere implementing it.

I copied the existing `*.witx` files into the `test-modules` directories
since they all assume that the `*.witx` has a particular shape. This
is a bit of duplication but allows the two to go out of sync so new
codegen tests don't have to also update the tests for linking.