Implement support for resources in wasmlink.
This commit implements support for resources and handle types in wasmlink.

A small runtime will be injected into the linked module that will be
responsible for tracking all currently-live resources within every submodule.

Each defined resource is given a specific identifier so that the runtime will
segment the tracking slabs per-resource.

To facilitate this feature, `AdaptedModule` was refactored such that it can
generate the adapter in coordination with a `Resources` helper type.

The `ModuleShim` helper was folded into `AdaptedModule`.

`CallAdapter` was updated to support adapting return values not coming from a
retptr (e.g. a single-value return of a handle).

We now fully rely on `Interface::all_bits_valid` to determine if we need to
adapt a type.  This means we will generate adapters for bools, flags, and enum
variants that are currently a no-op as we don't yet validate the value's
range against the underlying variant type. A future change will make these
adapter functions more useful.

A simple test case was added where resources are exchanged between two modules.