Copy test module witx files local to the implementation.
This commit copies the witx files local to the implementation for each test
module.

This frees up the tests/ directory to be about witx parsing only and the test
modules are free to have different test cases.

The `resource` and `resource-main` modules were renamed to `resources` and
`resources-main`, respectively, for consistency with the other test module
names.