Add file-based tests to wasmlink.
This commit adds file-based tests to wasmlink to make it easier to see the
generated code for adapted modules.

Additionally, `AdaptedModule` has been renamed to `ModuleAdapter` and exported
from the `wasmlink` crate.

The adapted module now conditionally imports and aliases a memory or the
canonical ABI functions relating to lists depending on whether the adapted
module needs them.