Fix send/sync bounds in imports
Some locations in the code used a different condition for determining
whether async should be used from other locations, and this mismatch
caused issues in the generated code. Instead use `None` as the only
indicator of "no async anywhere" and otherwise assume that everything
needs to be async unless explicitly mentioned otherwise.

Closes #41