Add a markdown code generator
This commit adds a new generation mode for Markdown which can be used to
generate both a `bindings.md` as well as a `bindings.html`, rendered by
`pulldown-cmark`, for documenting `*.witx` interfaces. This is currently
intended to somewhat closely mirror the markdown support in the upstream
WASI repository. Currently, though, it's quite simple and will likely
see more improvements over time!