rust-wasm: Use objects/methods for resources with methods
This is the same as the previous commit, except for the rust-wasm
generator. This enables imports and exports to use associated functions
and methods where appropriate to be more idiomatic within Rust. Note
that this only applies to resources with functions associated with them.

This change restructures how exports of a rust-wasm module are modeled.
Previously a trait was generated and a function was called to return a
singleton instance of the trait, but now the trait's methods no longer
take `&self` and it's expected that a zero-sized type is used to
implement the trait. Separate traits are generated for each resource
with methods as well.

Resource traits (for exports) take `&self` for methods, and otherwise
work with `Handle<T>` types for all other handle types (e.g. results or
other arguments).

cc #38