Use object-like and method-aware syntax in Rust-wasm & JS (#48)
* js: Use objects/methods for resources with methods

This commit generates more idiomatic JS when resources have methods and
static functions associated with them. Resources exported to JS now have
an object-like interface to remove the need to deal with the original
wasm object. Resources imported from JS have Typescript interfaces
generated which allows using method-call syntax to invoke methods. Note
that static methods on resources imported from JS are still freestanding
methods on the import object to prevent the generated glue code from
having to import a named type.

cc #38

* rust-wasm: Use objects/methods for resources with methods

This is the same as the previous commit, except for the rust-wasm
generator. This enables imports and exports to use associated functions
and methods where appropriate to be more idiomatic within Rust. Note
that this only applies to resources with functions associated with them.

This change restructures how exports of a rust-wasm module are modeled.
Previously a trait was generated and a function was called to return a
singleton instance of the trait, but now the trait's methods no longer
take `&self` and it's expected that a zero-sized type is used to
implement the trait. Separate traits are generated for each resource
with methods as well.

Resource traits (for exports) take `&self` for methods, and otherwise
work with `Handle<T>` types for all other handle types (e.g. results or
other arguments).

cc #38

* Fix & move demo

We've got two demos now, so make it clear that `crates/demo` is the
witx-bindgen demo. Additionally move the build script into that
directory.

* Fix wasmlink demo