Add missing exports to adapted modules.
This commit exports the memory and canonical ABI functions relating to lists to
the adapted module.

Without these exports, the adapted module cannot be used by another adapted
module.

A test was added to ensure that we can adapt a simple graph of imported modules
(`main -> b -> a`, where `b` and `a` are both adapted).

Fixes #51.