Initial commit of a wasmtime-py bindings generator (#55)
* Initial commit of a wasmtime-py bindings generator

Currently supports everything necessary to run `host.witx`, but there
are known missing holes:

* Endianness isn't handled
* Probably not as efficient as it could be
* Legacy witx features aren't supported

This generator is intended for host embeddings of Python using the
`wasmtime` package on PyPI. This will generate Python with annotated
types for usage with `mypy` and other type checkers as well (and `mypy`
is a bit part of the tests for all of this). The idioms generated here
are a sort of blend of JS idioms where ownership doesn't matter and
Wasmtime idioms where types like the `Store` must be passed around. This
is intended to feel like idiomatic Python, but being someone who hasn't
written a lot of Python I'm sure there's room for improvement here and
there, so this is just a first pass.

* Python syntax highlighting in demo

* Document markdown in readme

* Install Python on GHA