Expand the `Generator` trait so it can abstract the SpiderMonkey backend
This commit renames

* `preprocess` to `preprocess_one`,
* `generate` to `generate_one`, and
* `finish` to `finish_one`.

It also adds three new methods:

* `preprocess_all`,
* `generate_all`, and
* `finish_all`.

The `preprocess_all` method is called at the start of bindings generation, and
gives the generator a chance to inspect all imported and exported interfaces at
once. This is necessary for the SpiderMonkey backend so that it can assign
function indices in the generated glue code Wasm module for each imported and
exported function.

The `finish_all` method is called after all imported and exported interfaces
have been processed. It is an opportunity to generate a final file once the
generator has processed everything. This is necessary because generating a
single Wasm glue module for all imported and exported interfaces can't happen
until after we've processed each interface.

Finally, the `generate_all` method is essentially the main, top-level bindings
generation driver. It takes all imported and exported interfaces, calls
`preprocess_all` on them, and then loops over each interface, calling
`generate_one` on each of them with the appropriate import/export
direction. Finally, it calls `finish_all`.