Move spidermonkey tests to `tests/runtime/*`
This commit moves the spidermonkey tests into the new unit-test-like
directory of `tests/runtime`. This merges the import/export JS for now
but they can be split in the future if necessary. The only host which
supports these at this time is the Wasmtime host which uses a new helper
to compile and link in `spidermonkey.wasm` as well.