Implement async witx functions for a Wasmtime host
This commit is the initial implementation of `async function()` in witx
for the Wasmtime host runtime. This integration is unfortunately not as
nice as the previous JS implementation, but the general goal is to hook
up async witx functions to `async` functions in Rust in an at least
somewhat of an idiomatic fashion.

For exported functions not much has changed relative to the preexisting
`async` support for Wasmtime. Functions are still exposed as `async`
functions that take a store and the arguments as input. The major
consequence of this is that only one async method call on a wasm module
can be active at any point in time. The alternative to this design is to
embed the `Store<T>` within the bindings and have some sort of async
mutex around it, but that doesn't feel appropriate at this time since
it's not really how Wasmtime's idioms are designed (where you're
typically given a store rather than embedding it).

For imported functions, however, the implementation is significantly
different from the previous `async` support for Wasmtime. The support
no longer uses `async_trait` for the reason of being able to call
multiple imports simultaneously. This means that imports receive their
arguments and return `'static` futures. This does not map to any
supported mode of `async_trait`. This means that host bindings for these
sorts of functions are likely going to be somewhat unidiomatic since
they'll require cloning state into host callbacks as necessary and/or
using things like `Rc<RefCell<_>>` or similar.

Overall I'm not super happy with how this ended up. I feel that it's
generally pretty un-ergonomic and confusing to use. Unfortunately though
I don't really know how to make it better at this time. On the plus side
it at least all works for the given use cases. There's almost surely a
bug or two with the generated code but this should be at least a
somewhat solid base to build on as we tweak things in the future.