Swap `import` and `export` in the Wasmtime bindings.
Swap the names of `import` and `export` in command-line arguments and macros
for the Wasmtime bindings. This makes them the opposite of the
`witx2::abi::Direction` argument to the underlying generator code.

As background, `Import` and `Export` in the generator code imply different
ABIs. In a wasm-calling-wasm use case, one wasm module would use the `Import`
ABI, the other would use the `Export` ABI, and there would be an adapter
layer between the two that translates from one ABI to the other.

With wasm-calling-host, we don't go through a separate adapter layer;
the binding code we generate on the host side just does everything
itself. So when the host is conceptually "exporting" a function to
wasm, it uses the `Import` ABI so that wasm can also use the `Import`
ABI and import it directly from the host.

The key is, from the user's perspective, these are all just
implementation details. For users, `export` means exporting functions
to be called, and `import` means importing functions to call, in both
wasm modules and host code.

This is also described in comments in crates/wasmtime-impl/src/lib.rs.