Generate a `name` section for spidermonkey wasm
This commit updates the generated wasm module for spidermonkey modules
to have a `name` section which contains debugging information about
functions, locals, memories, etc. This is not currently complete in that
everything is named, but it at least adds a lot of names to the output
to make it a bit more readable. My hope is that this could be expanded
as necessary over time to name more items to assist in debugging the
generated code as necessary.