Revert "Swap `import` and `export` in the Wasmtime bindings."
This reverts commit bd86bed02a84b595a2a56ca105d29e5f7e4f1605.

I've been doing work on the async bindings for Wasmtime and for the life
of me I can no longer keep import/export straight, so I'm proposing that
this change is reverted. The usage of import/export was only changed on
the surface for witx-bindgen tooling but it means that reading and
writing the internals of witx-bindgen many operations feel backwards and
I forget what is mapped where and which end swapped what.

This reopens the problem that the import/export terminology for the host
(you're not "importing" functionality but rather providing an import to
wasm) but I personally feel that in the long run there's got to be a
better solution than using import/export at all. For example profiles I
think are promising where it's one unit to talk about and profiles
themselves say what they require and what they provide. I'm hoping that
a refactoring in the future or changing how the tooling works will solve
the user-facing problem, but for now while witx-bindgen is under heavy
development I find myself frequently tripping up to these swapped terms
since internally witx-bindgen is now inconsistent.