Start an explainer document for the witx format
We don't actually have any documentation today of the `*.witx` format,
so this attempts to at least start a rudimentary document which explains
the syntax and some of its capabilities.

Overall there's not a ton of documentation for things in this repository
just yet so I don't know the best place to put this. For now I chose a
markdown file in the root of the repository. Presumably over time we'll
create some sort of book or something like that and we can move this
over or split it up at that time. In the meantime though the `*.witx`
format is relatively stable and probably worthwhile documenting as this
progresses along!