Further refactoring for `Direction`.
Make `Generator::preprocess_one` take a `Direction`, so that we preserve
the property that everything outside `Generator` thinks in terms of
developer-facing directions, and everything inside `Generator`
implementations thinks in terms of ABI variants.

And collect the various `Direction` enums so that we only have one,
declared in gen-core along with `Generator`.