Rename "witx" to "wai"
This commit renames this repository and everything internally to "wai"
instead of the old name "witx". This has been discussed at previous WASI
group meetings and this change is implementing the current consensus.
Functionally not much should be changed here but this is ripe for some
updates after this commit goes through.

For reference the term "wai" stands for "WebAssembly Interface"