Implement `AsFd` etc. for handle types in the rust-wasm bindings.
Implement `AsFd`/`FromFd`/`IntoFd` and `AsRawFd`/`FromRawFd`/`IntoRawFd`
traits for handle types in the rust-wasm bindings.

This makes it more convenient to implement types like `File` and
`TcpStream` as wrappers around handles, as they support the usual
I/O handle APIs, and can simply be constructed directly from `FromFd`
without having to use raw handles and the unsafe `from_raw_fd`.

These traits are part of the [I/O safety RFC], which are
[implemented in Rust nightly]. In order to avoid having wit-bindgen
depend on nightly, this PR uses the [io-lifetimes] crate, which
contains standalone definitions of the main types and traits. Once
I/O safety is stabilized, the code can be switched to use std.

[I/O safety RFC]: https://github.com/rust-lang/rfcs/blob/master/text/3128-io-safety.md
[implemented in Rust nightly]: https://github.com/rust-lang/rust/issues/87074
[io-lifetimes]: https://crates.io/crates/io-lifetimes