Require wit identifiers to be in `kebab-case`. (#119)
* Require interface identifiers to be in `kebab-case`.

Background: Identifiers in wit, such as function and type names, as well
as argument and return value names, are mapped into source languages
following the conventions of the source languages. For example, Rust uses
CamelCase for types and snake_case for functions, while other languages
have other conventions. To ensure that this works reliably, this PR
proposes to define and enforce a naming convention for identifiers in
wit. This is not yet an official proposal, and the specifics are likely
to evolve for a while as the details are worked out.

Specifically, this PR requires identifiers in wit to be `kebab-case`, as
opposed to `snake_case` or `CamelCase`. And, it requires names to be
composed of valid Unicode Identifiers, and it requires identifiers to be
in stream-safe NFC form.

It also requires wit source files to not contain any bidirectional
formatting characters, control codes (other than newline, CR, and tab), or
codepoints that Unicode strongly discourages using.