wit-component: update interface-only encodings.
This commit updates the `wit2wasm` and `wasm2wit` tools to encode and decode
direct exports rather than using an instance type to model the interface.

The intent behind this is to better align with a component that exports its
"default" interface as direct function exports rather than always exporting
instances.