fix(wasmtime): Fix borrowchecker errors in Wasmtime list lowering
Fixes code generation for Wasmtime list lowering.

List lowering happens in a for loop and thus opens a new scope, which
invalidates previous caller_memory bindings.

This previously was not ensured, resulting in borrow checker errors in
the generated code if the items in a list required allocation.

The `lists.wit` tests is modified to contain previously problematic definitions.

Fixes #187